refactor: organize presentation layer into views/ and hooks/ folders
Consistent structure across all features:
  presentation/
    views/    — *View.tsx files
    hooks/    — use*ViewModel.ts files
    actions/  — server actions (unchanged)
    utils/    — utilities (where needed)
