+++ src/app/(auth)/login/page.tsx
@@ -1,4 +1,4 @@
-import { LoginView } from '@/features/auth/presentation/LoginView';
+import { LoginView } from '@/features/auth/presentation/views/LoginView';
 
 export default function LoginPage() {
   return <LoginView />;
