+++ src/app/(main)/budget-settings/new/page.tsx
@@ -1,4 +1,4 @@
-import { BudgetSettingNewView } from '@/features/budget-settings/presentation/BudgetSettingNewView';
+import { BudgetSettingNewView } from '@/features/budget-settings/presentation/views/BudgetSettingNewView';
 
 export default function BudgetSettingNewPage() {
   return <BudgetSettingNewView />;
