+++ src/app/(main)/budget-settings/page.tsx
@@ -1,4 +1,4 @@
-import { BudgetSettingsView } from '@/features/budget-settings/presentation/BudgetSettingsView';
+import { BudgetSettingsView } from '@/features/budget-settings/presentation/views/BudgetSettingsView';
 
 export default function BudgetSettingsPage() {
   return <BudgetSettingsView />;
