+++ src/app/(main)/categories/page.tsx
@@ -1,4 +1,4 @@
-import { CategoriesView } from '@/features/categories/presentation/CategoriesView';
+import { CategoriesView } from '@/features/categories/presentation/views/CategoriesView';
 
 export default function CategoriesPage() {
   return <CategoriesView />;
