+++ src/app/(main)/dashboard/page.tsx
@@ -1,7 +1,7 @@
 import { redirect } from 'next/navigation';
 import { createSupabaseServerClient } from '@/lib/auth';
 import { createServerContainer } from '@/shared/di/container.server';
-import { DashboardView } from '@/features/dashboard/presentation/DashboardView';
+import { DashboardView } from '@/features/dashboard/presentation/views/DashboardView';
 
 export default async function DashboardPage() {
   const supabase = await createSupabaseServerClient();
