+++ src/app/(main)/settings/page.tsx
@@ -1,4 +1,4 @@
-import { SettingsView } from '@/features/settings/presentation/SettingsView';
+import { SettingsView } from '@/features/settings/presentation/views/SettingsView';
 
 export default function SettingsPage() {
   return <SettingsView />;
