+++ src/app/(main)/split-bill/[id]/page.tsx
@@ -1,4 +1,4 @@
-import { SplitBillManageView } from '@/features/split-bill/presentation/SplitBillManageView';
+import { SplitBillManageView } from '@/features/split-bill/presentation/views/SplitBillManageView';
 
 export default async function SplitBillManagePage({ params }: { params: Promise<{ id: string }> }) {
   const { id } = await params;
