+++ src/app/(main)/split-bill/new/page.tsx
@@ -1,5 +1,5 @@
 import { Suspense } from 'react';
-import { SplitBillNewView } from '@/features/split-bill/presentation/SplitBillNewView';
+import { SplitBillNewView } from '@/features/split-bill/presentation/views/SplitBillNewView';
 
 export default function SplitBillNewPage() {
   return (
