+++ src/app/(main)/split-bill/page.tsx
@@ -1,4 +1,4 @@
-import { SplitBillListView } from '@/features/split-bill/presentation/SplitBillListView';
+import { SplitBillListView } from '@/features/split-bill/presentation/views/SplitBillListView';
 
 export default function SplitBillPage() {
   return <SplitBillListView />;
