+++ src/app/(main)/transactions/new/page.tsx
@@ -1,4 +1,4 @@
-import { TransactionNewView } from '@/features/transactions/presentation/TransactionNewView';
+import { TransactionNewView } from '@/features/transactions/presentation/views/TransactionNewView';
 
 export default function TransactionNewPage() {
   return <TransactionNewView />;
