+++ src/app/(main)/transactions/page.tsx
@@ -1,4 +1,4 @@
-import { TransactionsView } from '@/features/transactions/presentation/TransactionsView';
+import { TransactionsView } from '@/features/transactions/presentation/views/TransactionsView';
 
 export default function TransactionsPage() {
   return <TransactionsView />;
