+++ src/app/setup/page.tsx
@@ -1,4 +1,4 @@
-import { SetupView } from '@/features/setup/presentation/SetupView';
+import { SetupView } from '@/features/setup/presentation/views/SetupView';
 
 export default function SetupPage() {
   return <SetupView />;
