+++ src/features/budget-settings/presentation/views/BudgetSettingEditView.tsx
@@ -5,8 +5,8 @@ import { useRouter } from 'next/navigation';
 import { Card, CardContent, CardHeader, CardTitle, Button, CurrencyInput } from '@handharr-labs/ui-xpnsio';
 import { ROUTES } from '@/shared/presentation/navigation/routes';
 import { CURRENCY_OPTIONS } from '@/shared/presentation/constants/currencyOptions';
-import { useBudgetSettingEditViewModel } from './useBudgetSettingEditViewModel';
-import type { EditableCategoryItem } from './useBudgetSettingEditViewModel';
+import { useBudgetSettingEditViewModel } from '../hooks/useBudgetSettingEditViewModel';
+import type { EditableCategoryItem } from '../hooks/useBudgetSettingEditViewModel';
 
 const COLOR_OPTIONS = [
   '#6366f1', '#ec4899', '#f59e0b', '#10b981', '#3b82f6',
