+++ src/features/categories/presentation/hooks/useCategoriesViewModel.ts
