+++ src/features/categories/presentation/views/CategoriesView.tsx
@@ -4,9 +4,9 @@ import { useState } from 'react';
 import { Plus, Layers, UtensilsCrossed, Car, Home, ShoppingBag, Heart, BookOpen, Tv, Plane, Circle } from 'lucide-react';
 import { Button, FormDialog, ItemGroupSection } from '@handharr-labs/ui-xpnsio';
 import type { IconOption, FormDialogState } from '@handharr-labs/ui-xpnsio';
-import { useCategoriesViewModel } from './useCategoriesViewModel';
+import { useCategoriesViewModel } from '../hooks/useCategoriesViewModel';
 import type { Category } from '@/features/categories/domain/entities/Category';
-import { getCategoryIcon } from './utils/getCategoryIcon';
+import { getCategoryIcon } from '../utils/getCategoryIcon';
 
 const MASTER_LABELS: Record<string, string> = {
   daily: 'Daily Spend',
