+++ src/features/trips/presentation/hooks/useTripDetailViewModel.ts
