+++ src/features/trips/presentation/hooks/useTripListViewModel.ts
