+++ src/features/trips/presentation/hooks/useTripState.ts
