+++ src/features/trips/presentation/views/TripListView.tsx
@@ -4,7 +4,7 @@ import { useState } from 'react';
 import { useRouter } from 'next/navigation';
 import { Plus, MapPin, X } from 'lucide-react';
 import { Button } from '@handharr-labs/ui-xpnsio';
-import { useTripListViewModel } from '../state/useTripListViewModel';
+import { useTripListViewModel } from '../hooks/useTripListViewModel';
 import { TripCardComponent } from './TripCardComponent';
 
 export function TripListView() {
